refactor: re-organize map header components
